+++ lib/src/wire/builders/widget_builders/text_builder.dart
@@ -7,9 +7,24 @@ class TextBuilder extends WidgetBuilder {
   material.Widget build(Element element) {
     return material.Text(
       element.data["text"],
-      style: element.data["style"] == null || element.data["style"]["data"] == null
+      style: element.data["style"] == null
           ? null
-          : TextStyle.fromJson(element.data["style"]["data"]).build(),
+          : TextStyle.fromJson(element.data["style"]).build(),
+      textAlign: element.data["textAlign"] == null
+          ? null
+          : TextAlign.fromJson(element.data["textAlign"]).build(),
+      textDirection: element.data["textDirection"] == null
+          ? null
+          : TextDirection.fromJson(element.data["textDirection"]).build(),
+      softWrap: element.data["softWrap"] ?? false,
+      overflow: element.data["overflow"] == null
+          ? null
+          : TextOverflow.fromJson(element.data["overflow"]).build(),
+      maxLines: element.data["maxLines"],
+      textScaleFactor: element.data["textScaleFactor"]?.toDouble(),
+      textWidthBasis: element.data["textWidthBasis"] == null
+          ? null
+          : TextWidthBasis.fromJson(element.data["textWidthBasis"]).build(),
     );
   }
 }